Fix gestures bugs
