+++ package.json
@@ -10,6 +10,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@use-gesture/react": "^10.3.1",
     "hammerjs": "^2.0.8",
     "lodash": "^4.17.21",
     "react": "^18.3.1",
